add config for taint info, add snip fallback
This adds a preference and configuration option for how
taint information icons are rendered. By default, Unicode
characters such as LOCK are used, but on some font setups
(for example, my Mac OS laptop) glyphs for those characters
are not available. Or if the glyphs are present, they may
not be monospaced. This option also makes it possible to
turn off the taint icons entirely.